# server/swagger/mastermind-v3.json
{
    "swagger": "2.0",
    "info": {
        "title": "MasterMind Middleware API",
        "description": "The data storage middleware for MasterMind",
        "version": "3.0.0"
    },
    "host": "localhost:3000",
    "schemes": [
        "http"
    ],
    "basePath": "/v3",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/projects": {
            "x-handler": "Projects",
            "get": {
                "summary": "Retrieve Projects",
                "description": "The Projects endpoint returns projects.\n",
                "operationId": "getProjects",
                "tags": [
                    "Projects"
                ],
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "A set of ids to which this set should be reduced.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "executiveSponsor",
                        "in": "query",
                        "description": "Filter projects by only those belonging to this executive sponsor (specified by the ID of the executive sponsor Person)",
                        "type": "string"
                    },
                    {
                        "name": "committed",
                        "in": "query",
                        "description": "Filter projects by only those that are committed",
                        "type": "boolean"
                    },
                    {
                        "name": "current",
                        "in": "query",
                        "description": "Filter projects by only those that are current",
                        "type": "boolean"
                    },
                    {
                        "name": "person",
                        "in": "query",
                        "description": "Filter projects to those to whom the person is assigned, where the person is provided by ID",
                        "type": "string"
                    },
                    {
                        "name": "states",
                        "in": "query",
                        "description": "Filter projects to those in one of the given states",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "planning",
                                "active",
                                "done",
                                "poc",
                                "preSales",
                                "supportActive",
                                "clientActive"
                            ]
                        }
                    },
                    {
                        "name": "types",
                        "in": "query",
                        "description": "Filter projects to those of one of the given types",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "invest",
                                "poc",
                                "paid"
                            ]
                        }
                    },
                    {
                        "name": "startingAfter",
                        "in": "query",
                        "description": "Filter projects to those starting on or after the given date.",
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "startingBefore",
                        "in": "query",
                        "description": "Filter projects to those starting before the given date.",
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "endingBefore",
                        "in": "query",
                        "description": "Filter projects to those ending before the given date.",
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "name": "endingAfter",
                        "in": "query",
                        "description": "Filter projects to those ending after the given date.",
                        "type": "string",
                        "format": "date"
                    }

                ],
                "responses": {
                    "200": {
                        "description": "An array of projects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Project"
                            }
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to view Projects",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Project",
                "description": "Create a new project.\nAn ID in the POST-ed project will be ignored.\n",
                "operationId": "createSingleProject",
                "tags": [
                    "Projects"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "body",
                        "description": "Data representing the project to create",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly created project",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to create new Projects",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "x-handler": "Projects",
            "get": {
                "summary": "Retrieve Project",
                "description": "The Project endpoint returns a single project.\n",
                "operationId": "getSingleProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the project to retrieve.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Projects"
                ],
                "responses": {
                    "200": {
                        "description": "The single project specified by the ID",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to view Projects",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A project with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Project",
                "description": "Updates an existing project.\n",
                "operationId": "updateSingleProject",
                "tags": [
                    "Projects"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the project to update.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "project",
                        "in": "body",
                        "description": "New data representing the project",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly updated project",
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to update Projects",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A project with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Project",
                "description": "Deletes a single Project.\n",
                "operationId": "deleteSingleProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the project to delete.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Projects"
                ],
                "responses": {
                    "200": {
                        "description": "The single project specified by the ID was deleted"
                    },
                    "403": {
                        "description": "The current user is not allowed to delete Projects",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A project with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Project": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "committed": {
                    "type": "boolean"
                },
                "created": {
                    "type": "object",
                    "properties": {
                        "date": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "by": {
                            "type": "string",
                            "description": "ID of Person"
                        }
                    }
                },
                "client": {
                    "type": "string",
                    "description": "ID of Client"
                },
                "description": {
                    "type": "string"
                },
                "primaryContact": {
                    "type": "string"
                },
                "executiveSponsor": {
                    "type": "string",
                    "description": "ID of Person"
                },
                "initStartDate": {
                    "type": "string",
                    "format": "date"
                },
                "initEndDate": {
                    "type": "string",
                    "format": "date"
                },
                "modified": {
                    "type": "object",
                    "properties": {
                        "date": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "by": {
                            "type": "string",
                            "description": "ID of Person"
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "salesSponsor": {
                    "type": "string",
                    "description": "ID of Person"
                },
                "startDate": {
                    "type": "string",
                    "format": "date"
                },
                "endDate": {
                    "type": "string",
                    "format": "date"
                },
                "state": {
                    "default": "planning",
                    "enum": [
                        "planning",
                        "active",
                        "done",
                        "poc",
                        "preSales",
                        "supportActive",
                        "clientActive"
                    ]
                },
                "type": {
                    "enum": [
                        "invest",
                        "poc",
                        "paid"
                    ]
                },
                "terms": {
                    "type": "object",
                    "properties": {
                        "fixedBidServicesRevenue": {
                            "type": "number"
                        },
                        "includesProjectManagementOverhead": {
                            "type": "boolean"
                        },
                        "billingFrequency": {
                            "type": "string"
                        },
                        "billingDate": {
                            "type": "string"
                        },
                        "type": {
                            "type": "string"
                        },
                        "softwareEstimate": {
                            "type": "number"
                        }
                    }
                }
            },
            "required": [
                "client",
                "name",
                "type",
                "startDate",
                "state",
                "executiveSponsor"
            ]
        },
...
        "Error": {
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "oauth2": {
            "type": "oauth2",
            "authorizationUrl": "http://localhost:3000/auth/google",
            "flow": "implicit",
            "scopes": {
                "https://www.googleapis.com/auth/plus.login": "authenticated via Google Plus"
            }
        }
    },
    "security": [
        {
            "oauth2": [
                "https://www.googleapis.com/auth/plus.login"
            ]
        }
    ]
}